Create `SettingsManager` and move all data settings data to `settings` package.
